+++ src/Spice86.Core/Emulator/CPU/CfgCpu/ParsedInstruction/Instructions/Grp4Callback.cs
@@ -6,7 +6,7 @@ using Spice86.Core.Emulator.CPU.CfgCpu.ParsedInstruction.Prefix;
 using Spice86.Shared.Emulator.Memory;
 
 public class Grp4Callback : InstructionWithModRm {
-    public Grp4Callback(SegmentedAddress address, InstructionField<byte> opcodeField, List<InstructionPrefix> prefixes,
+    public Grp4Callback(SegmentedAddress address, InstructionField<ushort> opcodeField, List<InstructionPrefix> prefixes,
         ModRmContext modRmContext, InstructionField<byte> callbackNumber) : base(address, opcodeField, prefixes,
         modRmContext) {
         CallbackNumber = callbackNumber;
