+++ src/Spice86.Core/Emulator/CPU/CfgCpu/ParsedInstruction/Instructions/Grp5RmCallNear.cs
@@ -6,7 +6,7 @@ using Spice86.Core.Emulator.CPU.CfgCpu.ParsedInstruction.Prefix;
 using Spice86.Shared.Emulator.Memory;
 
 public class Grp5RmCallNear : InstructionWithModRm {
-    public Grp5RmCallNear(SegmentedAddress address, InstructionField<byte> opcodeField,
+    public Grp5RmCallNear(SegmentedAddress address, InstructionField<ushort> opcodeField,
         List<InstructionPrefix> prefixes, ModRmContext modRmContext) : base(address, opcodeField, prefixes,
         modRmContext) {
     }
